+++ src/routes/(signin)/+layout.svelte
@@ -1,0 +1,9 @@
+<div class="bg w-screen h-screen flex flex-row justify-start items-center">
+	<slot />
+</div>
+
+<style>
+	.bg {
+		background-image: url($lib/assets/images/login-background.jpg);
+	}
+</style>
