+++ src/routes/(signin)/+layout.ts
@@ -1,0 +1,5 @@
+export function load({ url }) {
+	return {
+		url: url.pathname
+	};
+}
